backend1/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
email-validator==2.2.0
bcrypt==4.2.0
python-dotenv==1.0.1
stripe==11.5.0
PyMySQL==1.1.1
